+++ src/eval-frame/actions/actions.js
@@ -161,7 +161,7 @@ function evaluateCode(code, language, state, evalId) {
       );
     };
 
-    return ensureLanguageAvailable(language, state, dispatch)
+    return ensureLanguageAvailable(language, state, evalId, dispatch)
       .then(languageEvaluator =>
         runCodeWithLanguage(languageEvaluator, code, messageCallback)
       )