+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - move reps to react inspector
 - change nomenclature around user output reps _including in API_ (**breaking change**)
 - removed errant "0" in landing page
+- plugins now throw network errors to the iodide console and catch syntax errors within plugin loaders
 
 # 0.0.3 (2019-02-04)