+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - move reps to react inspector
 - change nomenclature around user output reps _including in API_ (**breaking change**)
 - removed errant "0" in landing page
+- implements a language selector menu on the right side of the console input
 
 # 0.0.3 (2019-02-04)